requirements.txt
numpy
python-binance
backtrader
TA-Lib
websocket
datetime
websocket-client
pandas
psycopg2
tensorflow
matplotlib